decks of 52 cards

Functions that extract information from a card

card_list()

card_suit(card = card_list())

card_rank(card = card_list())

card_style(card = card_list(), four_color = TRUE)

card_color(card = card_list(), four_color = TRUE)

card_html(card = card_list(), size = 172, four_color = TRUE)

card_specials()

card_htmlhex(card = card_list())

card_unicode(card = card_list())

Arguments

card
a card. A list of availale cards can be found with card_list()

four_color
should a four color deck be applied. This means that each suit will have it's own color (black, red, blue and green). Otherwise, use two colors (black and red).

size
the font-size to use

Examples

# show all valid cards
card_list()

#>  [1] "2s" "2h" "2c" "2d" "3s" "3h" "3c" "3d" "4s" "4h" "4c" "4d" "5s" "5h" "5c"
#> [16] "5d" "6s" "6h" "6c" "6d" "7s" "7h" "7c" "7d" "8s" "8h" "8c" "8d" "9s" "9h"
#> [31] "9c" "9d" "Ts" "Th" "Tc" "Td" "Js" "Jh" "Jc" "Jd" "Qs" "Qh" "Qc" "Qd" "Ks"
#> [46] "Kh" "Kc" "Kd" "As" "Ah" "Ac" "Ad"

# determine rank and suit and color
card_rank("Jc")

#> [1] "J"

card_suit("7d")

#> [1] "d"

card_color("Ad")

#> [1] "blue"

card_color("Ad", four_color = FALSE)

#> [1] "red"

# style cards either in the console or via HTML
cat(card_unicode("5h"), "\n")

#> ğŸ‚µ 

cat(card_style("5h"), "\n")

#> ğŸ‚µ 

card_html("5h")

# iterate over several cards for styling
card_list() %>% sapply(card_style) %>% cat("\n")

#> ğŸ‚¢ ğŸ‚² ğŸƒ’ ğŸƒ‚ ğŸ‚£ ğŸ‚³ ğŸƒ“ ğŸƒƒ ğŸ‚¤ ğŸ‚´ ğŸƒ” ğŸƒ„ ğŸ‚¥ ğŸ‚µ ğŸƒ• ğŸƒ… ğŸ‚¦ ğŸ‚¶ ğŸƒ– ğŸƒ† ğŸ‚ ğŸ‚· ğŸƒ— ğŸƒ‡ ğŸ‚¨ ğŸ‚¸ ğŸƒ˜ ğŸƒˆ ğŸ‚© ğŸ‚¹ ğŸƒ™ ğŸƒ‰ ğŸ‚ª ğŸ‚º ğŸƒš ğŸƒŠ ğŸ‚« ğŸ‚» ğŸƒ› ğŸƒ‹ ğŸ‚­ ğŸ‚½ ğŸƒ ğŸƒ ğŸ‚® ğŸ‚¾ ğŸƒ ğŸƒ ğŸ‚¡ ğŸ‚± ğŸƒ‘ ğŸƒ 

card_list() %>% lapply(card_html, 70) %>% span() %>% browsable()